Bug fix Board drawing. Classes refactored.
