+++ lib/player.rb
@@ -1,0 +1,9 @@
+class Player
+  attr_reader :name
+  attr_accessor :type
+  def initialize(name)
+    @name = name
+    @score
+    @type
+  end
+end
